Initial Commit
This is the initial (semi) functioning notebook for examining light curves of supernovae in the TESS field using the eleanor library to retreive TESS data. Currently, the code uses eleanor.Source(), which only retreives data from the most recent sector that observed the object. This is not ideal for the supernovae being observed, as they have peak times that may not be fully visible without patching together sectors.

 The next step I hope to take is to modify the code to use eleanor.multi_sectors() instead and implement it in a way that will allow me to specify which sectors I want to analyze and have it default to all sectors the target is observed. Of the ten supernovae in the notebook, about six have yet to be viewed by TESS, though two of those are returning a different error from the others that I plan to investigate.

 SN 2020 bpi is the supernova with the most potential for gaining meaningful results once multiple sectors are observed, and I am attempting to replicate the light curve of ASASSN-18tb to verify that the code is working as intended.

Another area to pursue is to possibly eliminate the background subtraction done by eleanor, as it is geared towards much shorter time scales and may remove some real data from supernovae when determining the aperture.