eleanor bug hunting
Currently, there seems to be an issue with eleanor and how it identifies objects and apertures. It appears that it isn't finding the right targets regardless of how the target is specified in the arguments for eleanor.Source(). Additionally, eleanor sets the aperture to be in the upper left hand corner for every object that has been tested. These issues are persistent across kernel restarts, jupyter relaunches, and system reboots.

The two notebooks to focus on is the one specifically for ASASSN-18tb and the sanity check notebook. The former is where the issue first became evident and the latter is where I identified the issue and explained it in more detail.

The ASASSN-18tb notebook has been fairly well-documented so that  the utility of the functions I've defined and the logical flow of the work are hopefully clear to anyone looking through it. There is a flag at the start of the document that should attempt to install the packages that may not be already on a system, but I have also listed the github repositories for each as a comment in case that cell doesn't run as expected.