functioning version of decoupled model (rough)
decoupled model (without Gaussian) is functional but rough; functions are using normalized data points, but only for the decoupled model. Should likely rewrite so all models use normalized flux. Also planning to remove plot_lc_model() and replace it with plot_lc_guesses() (pending name change) with an argument that determines if it bothers to run emcee or simply uses the guessed values to plot the given model. Also added improved lightcurves from Fausnaugh, but have not yet implemented them in the code.