plot_lc_guesses() changes
All models are now plotted with flux normalized to 40% of the peak light of TESS. No additional models have been added yet, but next is to make a decoupled Gaussian. plot_lc_guesses() has been improved in several ways and has made plot_lc_model() redundant, so the latter has been removed and the former has been renamed. Currently, the plots for it have some rough default coloring, but there's an argument to change them. Plotting without defining the number of models to plot may also be resource intensive in the case of a large number of steps and walkers, so use with caution.Yet to be added is the ability to save the plots in the functions themselves, but parts of it have been added and it should be fairly simple.