Decoupled Gaussian & Updated Data (wip)
I've added a decoupled model that includes a Gaussian. It seems to have two potential spots where the Gaussian might be fitted, but I need to investigate this further. It seems the peak of the gaussian is't particularly well constrained, as the factor is bumping up against the requirement that it be positive.

I also processed the new data from Dr. Fausnaugh and added the corrected frames to the mcmc notebook, though I haven't yet modified the mcmc to accomodate them and therefore have not yet run any fits for them. Still need to add the plot saving functionality as well and also make it so the errors are plotted for the data points