mcmc corrections
Due to some confusing regarding the least_squares implementation when correcting the lightcurves, I've switched to using an mcmc fitter for the corrections. This is in "MCMC Lightcurve Correction.ipynb" which is still a work in progress. It seems that the difference in corrections between the old and updated data sets are caused by the changes in bkg_model. I will investigate this further