optimization of mcmc
made optimizations to mcmc functions for fitting lightcurve, resulting in runs taking roughly 1/3 the time to execute. Added arguments to functions so the plots could be saved to a folder in the directory for use in the presentation and paper. Have determined that issue with sudden jump in flux is due to inherent nature of Gaussian, the only way to improve this would be to recreate the model described in the Kasen paper and used in Dr. Fausnaugh's recent paper. I've reached out to see if that code is available and am waiting to hear back