Model Checks
Added a few models (some non-functional currently) to attempt to consolidate errors. It seems that fitting TESS to ZTF is driving the power of the lightcurve fit to ~1.2 despite the fact that (a) the ZTF fits to roughly a power of 2 and (b) the raw TESS data fits to a power of ~1.8. This has some effect on the explosion time, which seems to lie outside of one standard deviation.